requirements.txt
numpy==1.26.4
pandas==2.2.0
pillow==10.2.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
tzdata==2024.1
